requirements.txt
python-telegram-bot>=21.6,<22.0
aiohttp>=3.10.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
